Rename the function into `StartAsyncPropagation`, it's clearer what it does. Move it in the `database` package. Because otherwise we have a cyclic dependence error when calling it from the `database` package (database -> service -> auth -> database). This might be solvable in another way though, but it would require refactoring.
