+++ app/database/propagation.go
@@ -1,19 +1,18 @@
-package service
+package database
 
 import (
 	"net/http"
 	"strings"
 	"time"
 
-	"github.com/France-ioi/AlgoreaBackend/app/database"
 	"github.com/France-ioi/AlgoreaBackend/app/logging"
 )
 
 const endpointTimeout = 3 * time.Second
 
-// SchedulePropagation schedules asynchronous propagation of the given types.
+// StartAsyncPropagation schedules asynchronous propagation of the given types.
 // If endpoint is an empty string, it will be done synchronously.
-func SchedulePropagation(store *database.DataStore, endpoint string, types []string) {
+func StartAsyncPropagation(store *DataStore, endpoint string, types []string) {
 	// Must not be called in a transaction because it calls an endpoint, which can take a long time.
 	store.MustNotBeInTransaction()
 
@@ -46,13 +45,13 @@ func SchedulePropagation(store *database.DataStore, endpoint string, types []str
 	}
 
 	if endpoint == "" || endpointFailed {
-		err := store.InTransaction(func(store *database.DataStore) error {
+		err := store.InTransaction(func(store *DataStore) error {
 			store.ScheduleItemsAncestorsPropagation()
 			store.SchedulePermissionsPropagation()
 			store.ScheduleResultsPropagation()
 
 			return nil
 		})
-		MustNotBeError(err)
+		mustNotBeError(err)
 	}
 }
